move inference code and fix all warnings
