+++ tools/webui/__init__.py
@@ -1,0 +1,10 @@
+import html
+
+
+def build_html_error_message(error):
+    return f"""
+    <div style="color: red; 
+    font-weight: bold;">
+        {html.escape(str(error))}
+    </div>
+    """
